Refactor tests - create utils
