+++ tests/actor/test_actor_manager.py
@@ -25,8 +25,7 @@ from tests.actor.fake_actor_classes import (
     FakeSimpleTimerActor,
 )
 
-def _run(coro):
-    return asyncio.get_event_loop().run_until_complete(coro)
+from tests.actor.utils import _run
 
 class ActorManagerTests(unittest.TestCase):
     def setUp(self):
