+++ tests/actor/test_method_dispatcher.py
@@ -18,9 +18,7 @@ from tests.actor.fake_actor_classes import (
     FakeSimpleActor
 )
 
-def _run(coro):
-    return asyncio.get_event_loop().run_until_complete(coro)
-
+from tests.actor.utils import _run
 
 class ActorMethodDispatcherTests(unittest.TestCase):
     def setUp(self):
