+++ tests/actor/test_state_manager.py
@@ -21,20 +21,12 @@ from dapr.serializers import DefaultJSONSerializer
 
 from tests.actor.fake_actor_classes import FakeSimpleActor
 
-
-def _async_mock(*args, **kwargs):
-    m = mock.MagicMock(*args, **kwargs)
-
-    async def mock_coro(*args, **kwargs):
-        return m(*args, **kwargs)
-
-    mock_coro.mock = m
-    return mock_coro
-
-def _run(coro):
-    return asyncio.get_event_loop().run_until_complete(coro)
-
 from tests.actor.fake_actor_classes import FakeDaprActorClient
+
+from tests.actor.utils import (
+    _async_mock,
+    _run
+)
 
 class ActorStateManagerTests(unittest.TestCase):
     def setUp(self):
@@ -48,8 +40,8 @@ class ActorStateManagerTests(unittest.TestCase):
             self._test_type_info, self._serializer, self._serializer, self._fake_client)
         self._fake_actor = FakeSimpleActor(self._runtime_ctx, self._test_actor_id)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock(return_value=base64.b64encode(b'"value1"')))
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.save_state_transactionally', new=async_mock())
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock(return_value=base64.b64encode(b'"value1"')))
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.save_state_transactionally', new=_async_mock())
     def test_add_state(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -66,7 +58,7 @@ class ActorStateManagerTests(unittest.TestCase):
         added = _run(state_manager.try_add_state('state1', 'value1'))
         self.assertFalse(added)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock())
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
     def test_get_state_for_no_state(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -80,7 +72,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertFalse(has_value)
         self.assertIsNone(val)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock(return_value=b'"value1"'))
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock(return_value=b'"value1"'))
     def test_get_state_for_existing_value(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -88,7 +80,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertTrue(has_value)
         self.assertEqual("value1", val)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock(return_value=b'"value1"'))
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock(return_value=b'"value1"'))
     def test_get_state_for_removed_value(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -102,7 +94,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertFalse(has_value)
         self.assertIsNone(val)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock())
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
     def test_set_state_for_new_state(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -112,7 +104,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertEqual(StateChangeKind.add, state.change_kind)
         self.assertEqual('value1', state.value)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock())
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
     def test_set_state_for_existing_state_only_in_mem(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -127,7 +119,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertEqual(StateChangeKind.add, state.change_kind)
         self.assertEqual('value2', state.value)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock(return_value=b'"value1"'))
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock(return_value=b'"value1"'))
     def test_set_state_for_existing_state(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -136,14 +128,14 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertEqual(StateChangeKind.update, state.change_kind)
         self.assertEqual('value2', state.value)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock())
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
     def test_remove_state_for_non_existing_state(self):
 
         state_manager = ActorStateManager(self._fake_actor)
         removed = _run(state_manager.try_remove_state('state1'))
         self.assertFalse(removed)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock(return_value=b'"value1"'))
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock(return_value=b'"value1"'))
     def test_remove_state_for_existing_state(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -153,7 +145,7 @@ class ActorStateManagerTests(unittest.TestCase):
         state = state_manager._state_change_tracker['state1']
         self.assertEqual(StateChangeKind.remove, state.change_kind)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock())
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
     def test_remove_state_for_existing_state_in_mem(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -161,7 +153,7 @@ class ActorStateManagerTests(unittest.TestCase):
         removed = _run(state_manager.try_remove_state('state1'))
         self.assertTrue(removed)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock())
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
     def test_remove_state_twice_for_existing_state_in_mem(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -171,7 +163,7 @@ class ActorStateManagerTests(unittest.TestCase):
         removed = _run(state_manager.try_remove_state('state1'))
         self.assertFalse(removed)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock())
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
     def test_contains_state_for_removed_state(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -180,21 +172,21 @@ class ActorStateManagerTests(unittest.TestCase):
         exist = _run(state_manager.contains_state('state1'))
         self.assertTrue(exist)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock(return_value=b'"value1"'))
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock(return_value=b'"value1"'))
     def test_contains_state_for_existing_state(self):
 
         state_manager = ActorStateManager(self._fake_actor)
         exist = _run(state_manager.contains_state('state1'))
         self.assertTrue(exist)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock(return_value=b'"value1"'))
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock(return_value=b'"value1"'))
     def test_get_or_add_state_for_existing_state(self):
 
         state_manager = ActorStateManager(self._fake_actor)
         val = _run(state_manager.get_or_add_state('state1', 'value2'))
         self.assertEqual('value1', val)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock())
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
     def test_get_or_add_state_for_non_existing_state(self):
         state_manager = ActorStateManager(self._fake_actor)
         val = _run(state_manager.get_or_add_state('state1', 'value2'))
@@ -205,7 +197,7 @@ class ActorStateManagerTests(unittest.TestCase):
 
         self._fake_client.get_state.mock.assert_called_once_with(self._test_type_info._name, self._test_actor_id.id, 'state1')
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock(return_value=b'"value1"'))
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock(return_value=b'"value1"'))
     def test_get_or_add_state_for_removed_state(self):
 
         state_manager = ActorStateManager(self._fake_actor)
@@ -218,7 +210,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self.assertEqual(StateChangeKind.update, state.change_kind)
         self.assertEqual('value2', val)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock())
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
     def test_add_or_update_state_for_new_state(self):
         """adds state if state does not exist."""
         def test_update_value(name, value):
@@ -230,7 +222,7 @@ class ActorStateManagerTests(unittest.TestCase):
         state = state_manager._state_change_tracker['state1']
         self.assertEqual(StateChangeKind.add, state.change_kind)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock(return_value=b'"value1"'))
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock(return_value=b'"value1"'))
     def test_add_or_update_state_for_state_in_storage(self):
         """updates state value using update_value_factory if state is
         in the storage."""
@@ -243,7 +235,7 @@ class ActorStateManagerTests(unittest.TestCase):
         state = state_manager._state_change_tracker['state1']
         self.assertEqual(StateChangeKind.update, state.change_kind)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock(return_value=b'"value1"'))
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock(return_value=b'"value1"'))
     def test_add_or_update_state_for_removed_state(self):
         """add state value if state was removed."""
         def test_update_value(name, value):
@@ -255,7 +247,7 @@ class ActorStateManagerTests(unittest.TestCase):
         val = _run(state_manager.add_or_update_state('state1', 'value1', test_update_value))
         self.assertEqual('value1', val)
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock(return_value=b'"value1"'))
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock(return_value=b'"value1"'))
     def test_add_or_update_state_for_none_state_key(self):
         """update state value for StateChangeKind.none state """
         def test_update_value(name, value):
@@ -275,7 +267,7 @@ class ActorStateManagerTests(unittest.TestCase):
         with self.assertRaises(AttributeError):
             _run(state_manager.add_or_update_state('state1', 'value1', None))
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock())
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock())
     def test_get_state_names(self):
         state_manager = ActorStateManager(self._fake_actor)
         _run(state_manager.set_state('state1', 'value1'))
@@ -288,7 +280,7 @@ class ActorStateManagerTests(unittest.TestCase):
         self._fake_client.get_state.mock.assert_any_call(self._test_type_info._name, self._test_actor_id.id, 'state2')
         self._fake_client.get_state.mock.assert_any_call(self._test_type_info._name, self._test_actor_id.id, 'state3')
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock(return_value=b'"value0"'))
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock(return_value=b'"value0"'))
     def test_clear_cache(self):
         state_manager = ActorStateManager(self._fake_actor)
         _run(state_manager.set_state('state1', 'value1'))
@@ -298,8 +290,8 @@ class ActorStateManagerTests(unittest.TestCase):
 
         self.assertEqual(0, len(state_manager._state_change_tracker))
 
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=async_mock(return_value=b'"value3"'))
-    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.save_state_transactionally', new=async_mock())
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.get_state', new=_async_mock(return_value=b'"value3"'))
+    @mock.patch('tests.actor.fake_actor_classes.FakeDaprActorClient.save_state_transactionally', new=_async_mock())
     def test_save_state(self):
         state_manager = ActorStateManager(self._fake_actor)
         # set states which are StateChangeKind.add
