+++ tests/actor/utils.py
@@ -1,0 +1,14 @@
+import asyncio
+from unittest import mock
+
+def _async_mock(*args, **kwargs):
+    m = mock.MagicMock(*args, **kwargs)
+
+    async def mock_coro(*args, **kwargs):
+        return m(*args, **kwargs)
+
+    mock_coro.mock = m
+    return mock_coro
+
+def _run(coro):
+    return asyncio.get_event_loop().run_until_complete(coro)
